feat: extract nodeSettingsForm component
